bonds work

+++ src/bdsim/pyBrown/box.py
@@ -616,7 +616,7 @@ class Box():
 
 	def _initialize_force(self):
 
-		self.Fex = np.array( self.inp["external_force"] )
+		self.Fex = np.array( self.inp["external_force"], dtype = np.float )
 		self.F = np.zeros( 3*len(self.mobile_beads) )
 		self.E = 0.0
 		self._handle_external_force_restricted_to_region()
